사이트 세팅 페이지, model 준비

+++ src/main/java/com/devfox/bbvape/repo/SettingRepository.java
@@ -3,6 +3,6 @@ package com.devfox.bbvape.repo;
 import com.devfox.bbvape.model.Setting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SettingRepository extends JpaRepository<Setting, Long> {
-    Setting findById(long id);
+public interface SettingRepository extends JpaRepository<Setting, Integer> {
+    Setting findById(int id);
 }
